Python 3.9 fix: Prevent nx_pydot errors in node_coords_from_json
This is the same issue which also effects ManageGroupRelationshipsPresenter.full_chronograph_func,
where for versions of networkx available on python 3.9 pydot does not correctly escape string properties which include :
which is present for chronological graphs in the contraction node attribute.

This fix removes the contraction node attribute before the DAG is provided to pydot.