Tests: add tests for more util methods (polygonfunc, ellipsefunc, node_coords_fromjson)
Due to inconsistent graphviz behaviour the exact placement of nodes in node_coords_fromjson cannot be relied upon.

This can partially be addressed by splitting the method to one which produces an svg and etracts data, and another which just extracts data from a provided svg.

This could be the same method, just with differnt types of argument provided (str vs graph object)