Fix: Fix saving and calibrating models after GroupRelationship Enum was added.
The enum was not hashable, and was not json serialisable. This makes it hashable by implementing and testing __hash__, but also ensures that the enums are re-converted to strings prior to storing in the Model.

It might not be worth using the Enum, as it adds complexity. Namespaced constants migth be a much simpler alternative